backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
httpx==0.28.1
supabase==2.11.0
python-dotenv==1.0.1
apscheduler==3.10.4
numpy==2.2.3
scikit-learn==1.6.1
yfinance==0.2.52
feedparser==6.0.11
google-genai==1.7.0
pydantic==2.10.5
